Move processing hook execution in the main thread.
Since the execution of the processing does not leave anymore in the main thread, some operations in FinProcessing implementations could lead to UB when trying to modify the main thread. it's particulary true in "plugin context". This commit levrage _pre/_post processing hooks as a mean to give FinProcessing implementation a safe way to interact with the main thread.