Add call to dm.check_normalization()
Call to `dm.check_normalization()` is added. This function generates a boolean var with the following outputs: True if normalization is faulty, False otherwise.

A comment has been added to the bottom of the abstract processing to mark the place where a window that displays instructions on how to improve the normalization should be added.